hterm 0.8.2.10: Fix another missing-rows bug
Fix a bug where we'd shift the rowNodes element up to offset hidden nodes
before the top fold, but never compensated by making the rowNodes element
taller.

The active ingredient in this CL is the part where the topFoldOffset
is added to the height of rowNodes.  I had to turn syncRowNodesTop_
into syncRowNodesDimensions_ so that multiple callers could trigger
the recalc.

I've also included a number of previously uncommitted changes to the dev
changelog.

BUG=none
TEST=test_harness.html, 64/64 tests passed.

Change-Id: I2b9890b3a9132061b7cb9fc7847c353b74f1b58f
Reviewed-on: https://gerrit.chromium.org/gerrit/31406
Reviewed-by: Marius Schilder <mschilder@google.com>
Commit-Ready: Robert Ginda <rginda@chromium.org>
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>