Don't get confused by speech synthesis commands during initialization.
If Chrome called the TTS "Stop" function while the speech synthesis
component extension was still in the middle of initialization, it was
sending this stop message to NaCl, killing the plug-in. Fix it so that
it doesn't pass along that message until initialization is complete.

BUG=chromium:166253
TEST=Press Ctrl+Alt+Z on startup, immediately press Tab several times.

Change-Id: Ie6996f08ada38f6eb646ca8b0a266894b22b6be1
Reviewed-on: https://gerrit.chromium.org/gerrit/39725
Reviewed-by: Zachary Kuznia <zork@chromium.org>
Commit-Ready: Dominic Mazzoni <dmazzoni@chromium.org>
Tested-by: Dominic Mazzoni <dmazzoni@chromium.org>